requirements.txt
asgiref==3.7.2
Django==4.2.11
django-filter==23.5
djangorestframework==3.14.0
drf-yasg==1.21.7
importlib-metadata==7.0.1
inflection==0.5.1
Markdown==3.5.2
packaging==23.2
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.10.0
uritemplate==4.1.1
zipp==3.17.0
